# docs/apifox/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "餐饮外卖订单系统 API",
    "description": "Spring Boot 2.7 + MyBatis-Plus + MySQL + Redis + JWT。管理端前缀 /admin，用户端前缀 /user；除登录接口外均需在 Header 携带 token。本文件由 tools/export_openapi.py 从 Controller 源码自动提取生成，与实现同源。",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地环境"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "token": {
        "type": "apiKey",
        "in": "header",
        "name": "token",
        "description": "登录接口返回的 JWT，放在请求头 token 字段"
      }
    },
    "schemas": {
      "LoginDTO": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string",
            "example": "zhangsan"
          },
          "password": {
            "type": "string",
            "example": "123456"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "token": []
    }
  ],
  "paths": {
...
    "/admin/employee/login": {
      "post": {
        "tags": [
          "员工管理"
        ],
        "summary": "login",
        "operationId": "EmployeeController_login",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDTO"
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/user/login": {
      "post": {
        "tags": [
          "用户登录"
        ],
        "summary": "login",
        "operationId": "UserLoginController_login",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDTO"
              }
            }
          }
        },
        "security": []
      }
    },
...
  }
}